Unsafe int cast from long for java 7 compatibility

+++ src/main/java/org/c02e/jpgpj/util/Util.java
@@ -58,8 +58,7 @@ public class Util {
    */
     public static int bestFileBufferSize(long fileSize, int maxFileBufferSize) {
         return (fileSize>=maxFileBufferSize) ?
-            maxFileBufferSize :
-            Math.toIntExact(fileSize);
+            maxFileBufferSize : (int) fileSize;
     }
 
 }
